backend/pyproject.toml
[project]
name = "docmind"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6", "uvicorn[standard]==0.34.0",
    "pydantic==2.10.5", "pydantic-settings==2.7.1",
    "sqlalchemy==2.0.36", "asyncpg==0.30.0", "alembic==1.14.1",
    "redis[hiredis]==5.2.1", "qdrant-client==1.12.1",
    "anthropic==0.42.0", "openai==1.58.1", "cohere==5.13.4", "tiktoken==0.8.0",
    "langgraph==0.2.60", "langchain-core==0.3.29", "mcp[cli]==1.3.0",
    "tree-sitter==0.23.2", "tree-sitter-python==0.23.6",
    "tree-sitter-javascript==0.23.1", "tree-sitter-typescript==0.23.2",
    "tree-sitter-go==0.23.4", "mistune==3.0.2",
    "httpx==0.28.1", "PyGithub==2.5.0",
    "slack-bolt==1.21.2", "twilio==9.4.0",
    "PyJWT==2.10.1", "cryptography==44.0.0",
    "structlog==24.4.0", "python-multipart==0.0.20",
]
[project.optional-dependencies]
dev = ["pytest==8.3.4","pytest-asyncio==0.25.0","pytest-cov==6.0.0","httpx==0.28.1","ruff==0.8.6"]
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
[tool.setuptools.packages.find]
where = ["."]
